% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHT.R
\name{getLogCLR}
\alias{getLogCLR}
\title{getLogCLR}
\usage{
getLogCLR(dat)
}
\arguments{
\item{dat}{n x p matrix of numerical variables. For microbiome data, n: sample size, p: number of taxa}
}
\value{
n x p matrix of the log-ratio transformed data.
}
\description{
Obtain the centered log ratio transformation of compositional data.
}
\details{
A pseudocount is added to all taxa if zero values exist. A pseudocount of 0.01 is added if "dat" is read counts (rarefied or not), 
and a pseudocount that is 0.01 times the smallest nonzero relative abundances is added if "dat" is already proportions.
}
\examples{
n = 10; p = 100
set.seed(1)
dat = matrix(rbinom(n*p, size = 1000,  0.005),n,p)
X = getLogCLR(dat)
}
