% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHT.R
\name{GHT}
\alias{GHT}
\title{Generalized Hoteling's test}
\usage{
GHT(X, u = 0, nsim = 1000, target = "equal", centered = F)
}
\arguments{
\item{X}{n x p matrix of numerical variables}

\item{u}{a vector of numerical variables indicating the true value of the mean}

\item{nsim}{number of permutations.}

\item{target}{target matrix for covariance estimate. "equal": diagonal matrix with equal diagonal element;"unequal", diagonal matrix with unequal diagonal element;"identity": identity matrix}
}
\value{
p value
}
\description{
The fuction tests whether the mean of X is differnt from u, a hypothesized population average. 
GHT replaces the sample covariance matrix in classical Hoteling's test with a shrinkage based (positve definite) covariance matrix.
Significance is evaluated via permuation. 
The method is designed for paired microbiome studies, in which X is the paired differences after log-ratio transformation. 
However, the method is equally applicable to other high dimensional settings.
}
\examples{
set.seed(1)
n=10; p=100
dat = matrix(rnorm(n*p),n,p)
test1 = GHT(dat)
# A test similar to paired microbiome data
set.seed(1)
dat1 = matrix(rbinom(n*p, size = 1000,  0.005),n,p)
dat2 = matrix(rbinom(n*p, size = 1000,  0.005),n,p)
X1 = getCLR(dat1); X2 = getCLR(dat2)
X = X1 - X2
test2 = GHT(X)
}
